Add files via upload
These are all the cleaned and summarise data files from my Google Data Analytics Professional certificate capstone project with the exception of the August 2020 dataset and 202009-divvy-tripdatacleaned.csv datasets.  

The August 2020 dataset was cleaned in R following the process listed in the Appendix of the final report with some of the code used in an unstable cloud environment.

202009-divvy-tripdatacleaned.csv for the September 2020 cleaned dataset once compressed into a zip file is still to large to upload to GItHub. I am happy to email it upon request.

The original Datasets can be sourced from the following URL: https://divvy-tripdata.s3.amazonaws.com/index.html

These have been collected by Motivate International Inc. under the following Licence (https://ride.divvybikes.com/data-license-agreement) for the purposes of doing practice cases studies. Within the report as part of the Google Data Analytics Professional Certificate the company has been called Cycltics as the data is being used to provide insight into a theoretical marketing strategy for the fictional company as part of this capstone project.